project/production/requirements.txt
Django==2.2.1
django-webpack-loader==0.6.0
djangorestframework==3.9.4
Pillow==5.3.0
psycopg2==2.7.5
pytz==2018.5
requests==2.20.0
